refactor: Move ExcelSheet type to own module
